start rewrite jsbn
